Replace usage of the deprecated SGI STL extensions (ext/hash*) with the semi-official TR1 ones (tr1/unordered*). This silences a huge number of deprecation warnings on GCC 4.3+ and prepares the tree for future compiler revisions. However, this will probably render opencog unbuildable if using not-too-old compilers (it apears that at least GCC 3.X will break but I haven't been able to confirm this).

+++ src/dynamics/attention/ImportanceSpreadingAgent.cc
@@ -180,7 +180,7 @@ void ImportanceSpreadingAgent::spreadAtomImportance(Handle h)
 
     logger().fine("  +Hebbian links found %d", links->getSize());
 
-    totalDifference = sumTotalDifference(h, links);
+    totalDifference = static_cast<float>(sumTotalDifference(h, links));
     sourceSTI = a->getSTI(h);
 
     // if there is no hebbian links with > 0 weight
